requirements.txt
fastapi
uvicorn
crewai
groq
requests
pydantic
python-dotenv